@@ -7,7 +7,7 @@
   "repository": "csstools/postcss-env-function",
   "homepage": "https://github.com/csstools/postcss-env-function#readme",
   "bugs": "https://github.com/csstools/postcss-env-function/issues",
-  "main": "dist/index.js",
+  "main": "dist/index.cjs",
   "module": "dist/index.mjs",
   "files": [
     "dist"
@@ -32,15 +32,14 @@
     "postcss": "^8.3"
   },
   "devDependencies": {
-    "@babel/core": "7.15.5",
-    "@babel/eslint-parser": "7.15.4",
-    "@babel/preset-env": "7.15.6",
+    "@babel/core": "7.15.8",
+    "@babel/preset-env": "7.15.8",
     "@rollup/plugin-babel": "5.3.0",
-    "eslint": "7.32.0",
+    "eslint": "8.1.0",
     "postcss": "8.3.6",
     "postcss-tape": "6.0.1",
     "pre-commit": "1.2.2",
-    "rollup": "2.56.3"
+    "rollup": "2.58.3"
   },
   "babel": {
     "presets": [
@@ -54,11 +53,22 @@
   },
   "eslintConfig": {
     "env": {
-      "es6": true,
-      "node": true
+      "es6": true
     },
     "extends": "eslint:recommended",
-    "parser": "@babel/eslint-parser"
+    "parserOptions": {
+      "ecmaVersion": 12,
+      "sourceType": "module",
+      "ecmaFeatures": {
+        "modules": true
+      }
+    },
+    "rules": {
+      "semi": [
+        "error",
+        "never"
+      ]
+    }
   },
   "rollup": {
     "input": "src/index.js",
@@ -67,7 +77,8 @@
     ],
     "output": [
       {
-        "file": "dist/index.js",
+        "exports": "default",
+        "file": "dist/index.cjs",
         "format": "cjs"
       },
       {
