@@ -1,6 +1,6 @@
-import { parse } from 'postcss-values-parser';
-import updateEnvValue from './update-env-value';
-import walkEnvFuncs from './walk-env-funcs';
+import { parse } from 'postcss-values-parser'
+import updateEnvValue from './update-env-value'
+import walkEnvFuncs from './walk-env-funcs'
 
 /**
  * @param {string} originalValue
@@ -9,14 +9,14 @@ import walkEnvFuncs from './walk-env-funcs';
  */
 export default (originalValue, variables) => {
 	// get the ast of the original value
-	const ast = parse(originalValue);
+	const ast = parse(originalValue)
 
 	// walk all of the css env() functions
 	walkEnvFuncs(ast, node => {
 		// update the environment value for the css env() function
-		updateEnvValue(node, variables);
-	});
+		updateEnvValue(node, variables)
+	})
 
 	// return the stringified ast
-	return String(ast);
-};
+	return String(ast)
+}
