@@ -1,2 +1,2 @@
 // returns whether a node is a css env() function
-export default (node) => node && node.type === 'func' && node.name === 'env';
+export default (node) => node && node.type === 'func' && node.name === 'env'
